Rename "loginMenu.vue" to "LoginMenu.vue"
